Remove console logs, move getUserInfo out of thunks
